+++ au3/modules/import-export/RegisterImportPlugins.cpp
@@ -7,6 +7,7 @@
 
 #include "modules/import-export/mod-wavpack/ImportWavPack.h"
 #include "modules/import-export/mod-mpg123/ImportMP3_MPG123.h"
+#include "modules/import-export/mod-pcm/ImportPCM.h"
 
 void RegisterImportPlugins()
 {
@@ -18,5 +19,10 @@ void RegisterImportPlugins()
     static Importer::RegisteredImportPlugin sMP3ImportPlugin{
         "MP3",
         std::make_unique<MP3ImportPlugin>()
+    };
+
+    static Importer::RegisteredImportPlugin sPCMImportPlugin{
+        "PCM",
+        std::make_unique< PCMImportPlugin >()
     };
 }
