+++ src/au3wrap/CMakeLists.txt
@@ -562,6 +562,9 @@ set(AU3_SRC
     ${AU3_LIBRARIES}/lib-import-export/ExportUtils.cpp
     ${AU3_LIBRARIES}/lib-import-export/ExportUtils.h
 
+    ${AU3_LIBRARIES}/lib-import-export/GetAcidizerTags.cpp
+    ${AU3_LIBRARIES}/lib-import-export/GetAcidizerTags.h
+
     ${AU3_LIBRARIES}/lib-file-formats/AcidizerTags.h
     ${AU3_LIBRARIES}/lib-file-formats/FileFormats.cpp
     ${AU3_LIBRARIES}/lib-file-formats/FileFormats.h
@@ -595,6 +598,9 @@ set(AU3_SRC
 
     ${AU3_MODULES}/import-export/mod-pcm/ExportPCM.cpp
     ${AU3_MODULES}/import-export/mod-pcm/ExportPCM.h
+
+    ${AU3_MODULES}/import-export/mod-pcm/ImportPCM.cpp
+    ${AU3_MODULES}/import-export/mod-pcm/ImportPCM.h
 )
 
 if (AU_MODULE_EFFECTS_AUDIO_UNIT)
